Refactor RequestIterator into a mutable buffer
The standard-input buffer may trigger a notification to begin processing
at locations that do not correspond to fully qualified requests. That is
if two requests were sent simultaneously it might cut in the middle of
the header of the second. Which would leave one request that _could_ be
processed with another in an errored state. What is more, when the
buffer goes to read the next one it will likely fail.

The fix is to turn the RequestIterator into a mutable buffer to allow
partial requests to be resumed.