Rename the module JSONRPC to BaseProtocol
Using the nomenclature provided in the language server protocol as a
reference. It became clear that the protocol could be described as two
seperate but interacting modules, e.g., the base protocol and the
language server protocol.

The language server protocol extends the base protocol and adds the
higher order functionality. While the base protocol describes the format
of messages and how they are sent and received.

This enables some of the types currently in the LanguageServerProtocol
module, e.g., Header, HeaderField, etc.., to migrate to the new
BaseProtocol module where they more logically belong.

Finally, there was a bug in the old RequestIterator that prevented the
request buffer from ever emptying. This was caused by the fact that the
request buffer was a value type and was being copied. The solution was
to change the encapsulating type to a class so that it would get
reference semantics.