Add rules to Makefile for cleaning, installing, and uninstalling
This commit also removes the release build flag for using a static
stdlib (i.e. `-Xswiftc -static-stdlib`) because the stdlib is embedded
in libSwiftPM.dylib. After installing, having the stdlib in both of
those was causing issues and undefined behavior. I imagine that the
'undefined behavior' wouldn't be too terrible since the stdlib versions
would be the same, but it generated a lot of warning messages while
running the server.