FileHandle#waitForDataInBackgroundAndNotify is not implemented on Linux
The original run loop which is watching standard input for incoming text
does not work on Linux. The reason is that the function is marked
`NSUnimplemented()`. I did attempt a patch to swift-corelibs-foundation
to implement `waitForDataInBackgroundAndNotify` [1].

Unfortunately, there are other issues with the run loop as implemented
[2].

I felt the best solution was just switch to using DispatchIO to watch
standard input on a well defined run loop.

[1] https://github.com/apple/swift-corelibs-foundation/pull/1427
[2] https://github.com/RLovelett/vscode-swift/issues/36

Fixes issue #41